+++ src/Utils.ts
@@ -80,3 +80,6 @@ export function requireLocalFile(fileName: string): any {
         throw err;
     }
 }
+export function assertUnreachable(x: never): never {
+    throw new Error("Didn't expect to get here");
+}